+++ navigation/mtc_files/Markers/1b
@@ -1,0 +1,38 @@
+
+[Marker]
+FacetsCount=1
+Name=1b
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-13.977391161539257
+EndPos(0,1)=1.2440742880315269e-015
+EndPos(0,2)=-9.6300745155986073e-015
+EndPos(1,0)=13.977391161539257
+EndPos(1,1)=1.2515682934477468e-015
+EndPos(1,2)=-9.6789243286821155e-015
+
+[Facet1V2]
+EndPos(0,0)=13.977391161539257
+EndPos(0,1)=1.2515682934477468e-015
+EndPos(0,2)=-9.6789243286821155e-015
+EndPos(1,0)=14.019670530273789
+EndPos(1,1)=15.245637552880362
+EndPos(1,2)=-9.7055696812731187e-015
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=0.
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=0.
+R0,2=0.
+R1,2=0.
+R2,2=1.
