+++ navigation/mtc_files/Markers/2Coil
@@ -1,0 +1,39 @@
+
+[Marker]
+FacetsCount=1
+SliderControlledXpointsCount=0
+Name=2Coil
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-8.9266116684264851
+EndPos(0,1)=-7.9898502660127542e-016
+EndPos(0,2)=1.0214990704373533e-015
+EndPos(1,0)=8.9266116684264833
+EndPos(1,1)=-6.360897244681129e-016
+EndPos(1,2)=2.6438799470143262e-015
+
+[Facet1V2]
+EndPos(0,0)=8.9266116684264833
+EndPos(0,1)=-6.360897244681129e-016
+EndPos(0,2)=2.6438799470143262e-015
+EndPos(1,0)=9.0134937152069554
+EndPos(1,1)=-12.059011297429047
+EndPos(1,2)=3.6653790174516793e-015
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.20000000000000079
+R0,0=0.96592582628906842
+R1,0=-1.457167719820518e-016
+R2,0=-0.25881904510252007
+S1=-14.79999999999999
+R0,1=7.9979483404574392e-002
+R1,1=0.95105651629515409
+R2,1=0.29848749562898386
+S2=99.400000000000091
+R0,2=0.24615153938604106
+R1,2=-0.30901699437494573
+R2,2=0.91865005134999966
