+++ navigation/mtc_files/Markers/2b
@@ -1,0 +1,38 @@
+
+[Marker]
+FacetsCount=1
+Name=2b
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-13.9690559296461
+EndPos(0,1)=-9.08090028293946e-017
+EndPos(0,2)=3.7518297647387081e-015
+EndPos(1,0)=13.969055929646105
+EndPos(1,1)=-1.1456717217429434e-016
+EndPos(1,2)=3.7783785761971358e-015
+
+[Facet1V2]
+EndPos(0,0)=-13.924452046559736
+EndPos(0,1)=-15.196209648340083
+EndPos(0,2)=3.7349350665378906e-015
+EndPos(1,0)=-13.9690559296461
+EndPos(1,1)=-9.08090028293946e-017
+EndPos(1,2)=3.7518297647387081e-015
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=0.
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=0.
+R0,2=0.
+R1,2=0.
+R2,2=1.
