+++ navigation/mtc_files/Markers/3Coil
@@ -1,0 +1,39 @@
+
+[Marker]
+FacetsCount=1
+SliderControlledXpointsCount=0
+Name=3Coil
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-8.9434442564413246
+EndPos(0,1)=-6.1153268241647966e-016
+EndPos(0,2)=3.4944724709513123e-016
+EndPos(1,0)=8.9434442564413299
+EndPos(1,1)=-8.8089826871897666e-016
+EndPos(1,2)=-4.6592966279350834e-016
+
+[Facet1V2]
+EndPos(0,0)=-8.9439343674574907
+EndPos(0,1)=11.937904678416533
+EndPos(0,2)=-5.2417087064269685e-016
+EndPos(1,0)=-8.9434442564413246
+EndPos(1,1)=-6.1153268241647966e-016
+EndPos(1,2)=3.4944724709513123e-016
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.80000000000000182
+R0,0=0.9563047559630351
+R1,0=1.8041124150158794e-016
+R2,0=-0.29237170472273794
+S1=47.299999999999962
+R0,1=9.0347825433700193e-002
+R1,1=0.95105651629515331
+R2,1=0.29551442139416562
+S2=42.399999999999991
+R0,2=0.27806201495688238
+R1,2=-0.30901699437494834
+R2,2=0.90949986972269081
