+++ navigation/mtc_files/Markers/4Coil
@@ -1,0 +1,39 @@
+
+[Marker]
+FacetsCount=1
+SliderControlledXpointsCount=0
+Name=4Coil
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-8.9585812958876883
+EndPos(0,1)=-4.8569430893988515e-016
+EndPos(0,2)=2.3154142102162125e-015
+EndPos(1,0)=8.95858129588769
+EndPos(1,1)=-1.2210192124187057e-016
+EndPos(1,2)=4.6308284204324244e-016
+
+[Facet1V2]
+EndPos(0,0)=8.95858129588769
+EndPos(0,1)=-1.2210192124187057e-016
+EndPos(0,2)=4.6308284204324244e-016
+EndPos(1,0)=8.8914854228233384
+EndPos(1,1)=11.981845264775341
+EndPos(1,2)=3.4731213153243185e-016
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=-0.40000000000000141
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=10.700000000000001
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=110.09999999999999
+R0,2=0.
+R1,2=0.
+R2,2=1.
