+++ navigation/mtc_files/Markers/5Ref
@@ -1,0 +1,38 @@
+[Marker]
+FacetsCount=1
+SliderControlledXpointsCount=0
+Name=5Ref
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-8.9120676333490003
+EndPos(0,1)=2.3980817331903383e-016
+EndPos(0,2)=-1.2505552149377763e-015
+EndPos(1,0)=8.9120676333490003
+EndPos(1,1)=2.6645352591003756e-017
+EndPos(1,2)=0.
+
+[Facet1V2]
+EndPos(0,0)=-8.9951734202602971
+EndPos(0,1)=-12.045839883319603
+EndPos(0,2)=-4.5474735088646413e-016
+EndPos(1,0)=-8.9120676333490003
+EndPos(1,1)=2.3980817331903383e-016
+EndPos(1,2)=-1.2505552149377763e-015
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=0.
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=0.
+R0,2=0.
+R1,2=0.
+R2,2=1.
