+++ navigation/mtc_files/Markers/COOLCARD
@@ -1,0 +1,37 @@
+[Marker]
+FacetsCount=1
+Name=COOLCARD
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-31.763129600371474
+EndPos(0,1)=-6.4066430885648868e-016
+EndPos(0,2)=1.9712913472123418e-015
+EndPos(1,0)=31.763129600371467
+EndPos(1,1)=-6.4440270518456277e-016
+EndPos(1,2)=1.9656972001890368e-015
+
+[Facet1V2]
+EndPos(0,0)=-31.746690593113094
+EndPos(0,1)=25.423398067377672
+EndPos(0,2)=1.9712913472123418e-015
+EndPos(1,0)=-31.763129600371474
+EndPos(1,1)=-6.4066430885648868e-016
+EndPos(1,2)=1.9712913472123418e-015
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=0.
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=0.
+R0,2=0.
+R1,2=0.
+R2,2=1.
