+++ navigation/mtc_files/Markers/Pointer Template
@@ -1,0 +1,39 @@
+
+[Marker]
+FacetsCount=1
+SliderControlledXpointsCount=0
+Name=Pointer Template
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-8.9079792482800304
+EndPos(0,1)=4.7658354227811601e-016
+EndPos(0,2)=-3.8126683382249281e-015
+EndPos(1,0)=8.9079792482800322
+EndPos(1,1)=-3.754900636130611e-016
+EndPos(1,2)=4.6214161675453676e-016
+
+[Facet1V2]
+EndPos(0,0)=8.9079792482800322
+EndPos(0,1)=-3.754900636130611e-016
+EndPos(0,2)=4.6214161675453676e-016
+EndPos(1,0)=8.97037014055695
+EndPos(1,1)=-11.854493390526288
+EndPos(1,2)=-5.1990931884885378e-015
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=0.
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=0.
+R0,2=0.
+R1,2=0.
+R2,2=1.
