+++ navigation/mtc_files/Markers/TTblock
@@ -1,0 +1,37 @@
+[Marker]
+FacetsCount=1
+Name=TTblock
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-44.89806848621749
+EndPos(0,1)=1.5829351718288365e-017
+EndPos(0,2)=-2.4650420593630429e-015
+EndPos(1,0)=44.898068486217483
+EndPos(1,1)=8.5012292344588403e-016
+EndPos(1,2)=-2.4138677168217271e-015
+
+[Facet1V2]
+EndPos(0,0)=-0.25683199956524938
+EndPos(0,1)=42.807497943328485
+EndPos(0,2)=-8.1564930877711453e-003
+EndPos(1,0)=0.23846937714486499
+EndPos(1,1)=-42.932871913619174
+EndPos(1,2)=-8.1564930877712043e-003
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=0.
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=0.
+R0,2=0.
+R1,2=0.
+R2,2=1.
