+++ navigation/mtc_files/Markers/a
@@ -1,0 +1,37 @@
+[Marker]
+FacetsCount=1
+Name=a
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-32.971498168704073
+EndPos(0,1)=2.5755541490385065e-016
+EndPos(0,2)=1.6979881553090628e-016
+EndPos(1,0)=32.971498168704088
+EndPos(1,1)=3.2816886463184776e-016
+EndPos(1,2)=4.4082384801292978e-017
+
+[Facet1V2]
+EndPos(0,0)=-32.975185360462866
+EndPos(0,1)=46.723702252369826
+EndPos(0,2)=-8.9797450521152372e-018
+EndPos(1,0)=-32.971498168704073
+EndPos(1,1)=2.5755541490385065e-016
+EndPos(1,2)=1.6979881553090628e-016
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=0.
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=0.
+R0,2=0.
+R1,2=0.
+R2,2=1.
