+++ navigation/mtc_files/Markers/coil
@@ -1,0 +1,38 @@
+
+[Marker]
+FacetsCount=1
+Name=coil
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-70.299927206216339
+EndPos(0,1)=1.1368683772161603e-016
+EndPos(0,2)=1.3642420526593924e-015
+EndPos(1,0)=70.299927206216339
+EndPos(1,1)=-1.1368683772161603e-016
+EndPos(1,2)=6.8212102632969619e-016
+
+[Facet1V2]
+EndPos(0,0)=-69.651993385067129
+EndPos(0,1)=-65.744122964094188
+EndPos(0,2)=3.410605131648481e-016
+EndPos(1,0)=-70.299927206216339
+EndPos(1,1)=1.1368683772161603e-016
+EndPos(1,2)=1.3642420526593924e-015
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=0.
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=0.
+R0,2=0.
+R1,2=0.
+R2,2=1.
