+++ navigation/mtc_files/Markers/probe
@@ -1,0 +1,39 @@
+
+[Marker]
+FacetsCount=1
+SliderControlledXpointsCount=0
+Name=probe
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-61.602408326043872
+EndPos(0,1)=4.2184025144074082e-016
+EndPos(0,2)=-4.556586682229099e-016
+EndPos(1,0)=61.602408326043872
+EndPos(1,1)=-9.6115500328270063e-017
+EndPos(1,2)=-1.1391466705572748e-016
+
+[Facet1V2]
+EndPos(0,0)=-61.59657019736504
+EndPos(0,1)=11.949604571287503
+EndPos(0,2)=1.1391466705572748e-016
+EndPos(1,0)=-61.602408326043872
+EndPos(1,1)=4.2184025144074082e-016
+EndPos(1,2)=-4.556586682229099e-016
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=145.66478105979206
+R0,0=-8.4560706371216821e-002
+R1,0=-0.52109081782374
+R2,0=0.84930197604725266
+S1=-13.147426563284974
+R0,1=4.5008904130328986e-002
+R1,1=-0.85348236485506712
+R2,1=-0.5191743940434248
+S2=9.0630880541399712
+R0,2=0.99540126857813949
+R1,2=-5.6756022725552024e-003
+R2,2=9.5624798310161907e-002
