+++ navigation/mtc_files/Markers/sptr
@@ -1,0 +1,38 @@
+
+[Marker]
+FacetsCount=1
+Name=sptr
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-52.19791804784397
+EndPos(0,1)=1.2304185897405721e-016
+EndPos(0,2)=6.9721287355505385e-016
+EndPos(1,0)=52.19791804784397
+EndPos(1,1)=-7.4104692177650016e-017
+EndPos(1,2)=5.9481366254593747e-016
+
+[Facet1V2]
+EndPos(0,0)=-52.224959539439432
+EndPos(0,1)=11.532261461857525
+EndPos(0,2)=6.9918899867979109e-016
+EndPos(1,0)=-52.19791804784397
+EndPos(1,1)=1.2304185897405721e-016
+EndPos(1,2)=6.9721287355505385e-016
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=132.5915439593966
+R0,0=-3.3573683844187456e-002
+R1,0=-0.99175827123861893
+R2,0=0.12364602372465228
+S1=-3.8041069565099583
+R0,1=-3.9209531312008813e-002
+R1,1=-0.12231349340138305
+R2,1=-0.9917167045009595
+S2=-0.29142356239092942
+R0,2=0.99866682152128394
+R1,2=-3.8143685738751404e-002
+R2,2=-3.4779862432738673e-002
