+++ navigation/mtc_files/Markers/try4
@@ -1,0 +1,38 @@
+
+[Marker]
+FacetsCount=1
+Name=try4
+
+[Facet1]
+VectorsCount=2
+
+[Facet1V1]
+EndPos(0,0)=-57.030279071976651
+EndPos(0,1)=-1.7053025658242405e-016
+EndPos(0,2)=-3.1832314562052489e-015
+EndPos(1,0)=57.030279071976658
+EndPos(1,1)=1.7053025658242405e-016
+EndPos(1,2)=-1.8189894035458565e-015
+
+[Facet1V2]
+EndPos(0,0)=-56.190229598357341
+EndPos(0,1)=93.77396683857414
+EndPos(0,2)=0.
+EndPos(1,0)=-57.030279071976651
+EndPos(1,1)=-1.7053025658242405e-016
+EndPos(1,2)=-3.1832314562052489e-015
+
+[Tooltip2MarkerXf]
+Scale=1.
+S0=0.
+R0,0=1.
+R1,0=0.
+R2,0=0.
+S1=0.
+R0,1=0.
+R1,1=1.
+R2,1=0.
+S2=0.
+R0,2=0.
+R1,2=0.
+R2,2=1.
